% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{circle.esteem.for.clusters}
\alias{circle.esteem.for.clusters}
\title{Calculate the k-circle scores for clusters}
\usage{
circle.esteem.for.clusters(k)
}
\arguments{
\item{k}{a list of incidence matrices as produced by \link{k.circles}}
}
\value{
a list of data.frames
}
\description{
Calculate the k-circle scores for clusters
}
