% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{circle.esteem}
\alias{circle.esteem}
\title{Calculate the k-circle esteem scores}
\usage{
circle.esteem(k, esteem = esteem.graphs(k))
}
\arguments{
\item{k}{a list of incidence matrices as produced by \link{k.circles}}
}
\value{
a list of data.frames
}
\description{
Calculate the k-circle esteem scores
}
