% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{esteem.graphs}
\alias{esteem.graphs}
\title{Create graphs and adjacency matrixes weighted by circle esteem}
\usage{
esteem.graphs(k)
}
\arguments{
\item{k}{a list of incidence matrices as produced by \link{k.circles}}
}
\value{
list of data.frames
}
\description{
Create graphs and adjacency matrixes weighted by circle esteem
}
