% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{weighted.k.circle.projection.members}
\alias{weighted.k.circle.projection.members}
\title{Title}
\usage{
weighted.k.circle.projection.members(k, base = 10, sigma = 14)
}
\arguments{
\item{sigma}{}
}
\description{
Title
}
